Add files via upload
We developed a code for estimating phase in an experiment by creating a data which contains the phase value = pi/100 using quantum maximum likelihood estimation, which is a program in quantum machine learning for estimating a phase in a quantum experiment. Additionally, the sequential metrology scheme is heavily affected if thermodynamically consistent measurements and thermal initial states are taken into account. We estimated the phase using such realistic measurements and proved that one favors over other. We also evaluated the sensitivity of the measurement scheme.